+++ src/main/java/com/capgemini/ccsw/tutorial_server/load/LoadServiceImpl.java
@@ -114,7 +114,7 @@ public class LoadServiceImpl implements LoadService {
 		
 	}
 	@Override
-	public Integer validateLoan(Long game, String fecha) {
+	public Long validateLoan(Long game, String fecha) {
 		return this.loadRepository.validateLoan(game, fecha);
 	}
 	@Override
@@ -134,7 +134,7 @@ public class LoadServiceImpl implements LoadService {
 		return this.loadRepository.comprobarJuegos(game, fecha);
 	}
 	@Override
-	public Integer fechaInferior(Long client, String fecha) {
+	public Long fechaInferior(Long client, String fecha) {
 		return this.loadRepository.fechaInferior(client, fecha);
 	}
